+++ lang/en/coursecertificate.php
@@ -33,11 +33,9 @@ $string['automaticsenddisabled'] = 'The automatic sending of this certificate is
 $string['automaticsenddisabled_help'] = 'By leaving this disabled, students must click on the activity link displayed on the course page to receive the certificate, once they meet this activity\'s access restrictions.<br/><br/>
 By enabling it, students will automatically receive a PDF copy of the certificate once they meet this activity\'s access restrictions. Note that all students that already meet this activity\'s access restrictions will receive the certificate when enabling this.';
 $string['automaticsenddisabledalert'] = 'Students who meet this activity\'s access restrictions will be issued with their certificate once they access it.';
-$string['automaticsenddisabledinfo'] = 'Currently, {$a} students meet this activity\'s access restrictions and will be issued with their certificate once they access it.';
 $string['automaticsendenabled'] = 'The automatic sending of this certificate is enabled.';
 $string['automaticsendenabled_help'] = 'By leaving this enabled, students will automatically receive a PDF copy of the certificate once they meet this activity\'s access restrictions.<br/><br/>
 By disabling it, students will need to click on the activity link displayed on the course page to receive the certificate, once they meet this activity\'s access restrictions.';
-$string['certificateissues'] = 'Certificate issues';
 $string['certificatesarchived'] = 'Certificates archived';
 $string['certifiedusers'] = 'Certified users';
 $string['chooseatemplate'] = 'Choose a template...';
@@ -52,13 +50,13 @@ $string['courseurl'] = 'Course URL';
 $string['disableautomaticsend'] = 'Students will no longer automatically receive a PDF copy of the certificate as soon as they meet
  this activity\'s access restrictions. Instead, they will need to click on the activity link displayed on the course page to receive
  the certificate, once they meet this activity\'s access restrictions.';
-$string['enableautomaticsend'] = 'All students will automatically receive a PDF copy of the certificate as soon as they meet this activity\'s access restrictions.<br/><br/>
-Currently, {$a} students already meet these access restrictions but haven\'t accessed this activity yet. They will immediately receive their copy as well.<br/><br/>
-Students who have already accessed this activity will not receive the certificate again.';
 $string['enableautomaticsendpopup'] = 'All students will automatically receive a PDF copy of the certificate as soon as they meet this activity\'s access restrictions.<br/><br/>
 Students who already meet these access restrictions but haven\'t accessed this activity yet will immediately receive their copy as well.<br/><br/>
 Students who have already accessed this activity will not receive the certificate again.';
 $string['expirydate'] = 'Expiry date';
+$string['filterswarning'] = 'Filters defined in the activity context are only used for the activity title and description.';
+$string['filterswarninglist'] = 'Only filter(s) "{$a->list}" will be used when generating certificate PDF files and only if they are enabled on the <a href="{$a->link}">course level</a>.';
+$string['filterswarningnone'] = 'No filters will be used when generating certificate PDF files.';
 $string['issueddate'] = 'Date issued';
 $string['managetemplates'] = 'Manage certificate templates';
 $string['modulename'] = 'Course certificate';
@@ -81,10 +79,13 @@ $string['pluginname'] = 'Course certificate';
 $string['previewcoursefullname'] = 'Course full name';
 $string['previewcourseshortname'] = 'Course short name';
 $string['privacy:metadata'] = 'The course certificate activity does not store personal data.';
+$string['selecttemplatewarning'] = 'Once this activity issues at least one certificate, this field will be locked and will no longer be editable.';
+$string['taskissuecertificates'] = 'Issue course certificates';
+$string['template'] = 'Template';
+
+// Deprecated since 4.2.
+$string['certificateissues'] = 'Certificate issues';
 $string['revoke'] = 'Revoke';
 $string['revokeissue'] = 'Are you sure you want to revoke this certificate issue from this user?';
 $string['selectdate'] = 'Select date';
-$string['selecttemplatewarning'] = 'Once this activity issues at least one certificate, this field will be locked and will no longer be editable.';
 $string['status'] = 'Status';
-$string['taskissuecertificates'] = 'Issue course certificates';
-$string['template'] = 'Template';